controlplane: fix lint — move kubernetes-only metrics to a tagged file
golangci-lint runs with default build tags, so janitorIsLeaderGauge /
setJanitorIsLeader and workerHotIdlePersistFailuresCounter /
observeHotIdlePersistFailure — referenced only from kubernetes-tagged files
(janitor_leader_k8s.go, k8s_pool_lifecycle.go) — were flagged "unused" under the
default-tag analysis. Move them into worker_lifecycle_metrics_k8s.go
(//go:build kubernetes) so neither the symbol nor its sole caller is compiled in
the default build. hotIdleReapLastRunGauge / observeHotIdleReapRun stay in the
shared file (used by the untagged janitor.go + per_cp_hot_idle_reaper.go).

Verified: golangci-lint v2.11.4 (CI pin) reports 0 issues; builds clean on both
tags.
