+++ controlplane/worker_lifecycle_metrics.go
@@ -220,20 +220,15 @@ var workerHealthChecksCounter = promauto.NewCounterVec(prometheus.CounterOpts{
 	Help: "Worker RPC health-check probes partitioned by result (pass|fail) and image. Pass-rate complements the mark-lost transitions counter: a worker that's intermittently failing health checks but not yet crossing the consecutive-failure threshold is invisible without this.",
 }, []string{"result", "image"})
 
-var janitorIsLeaderGauge = promauto.NewGauge(prometheus.GaugeOpts{
-	Name: "duckgres_control_plane_janitor_is_leader",
-	Help: "1 if this control-plane replica currently holds the janitor leader lease, else 0. Summed across replicas it should be ~1; a sustained 0 means no replica is running the fleet-wide reapers (hot-idle TTL, orphan/stuck sweeps, headroom) — alert on it.",
-})
-
+// hotIdleReapLastRunGauge is updated by both the leader janitor (janitor.go,
+// untagged) and the per-CP fallback reaper (per_cp_hot_idle_reaper.go, untagged),
+// so it stays in this shared file. The kubernetes-only janitor_is_leader and
+// hot_idle_persist_failures metrics live in worker_lifecycle_metrics_k8s.go so
+// they are not flagged unused under the default-tag lint.
 var hotIdleReapLastRunGauge = promauto.NewGauge(prometheus.GaugeOpts{
 	Name: "duckgres_control_plane_hot_idle_reap_last_run_timestamp_seconds",
 	Help: "Unix timestamp of the most recent successful hot-idle reap pass on THIS replica — either the leader janitor or the per-CP fallback reaper. Alert when max() across replicas falls far behind now(): idle worker pods are not being retired and node capacity is leaking.",
 })
-
-var workerHotIdlePersistFailuresCounter = promauto.NewCounterVec(prometheus.CounterOpts{
-	Name: "duckgres_worker_hot_idle_persist_failures_total",
-	Help: "Failures to durably persist a worker's hot->hot_idle transition. The worker is held Hot (and stays reusable by its org) rather than advanced to an in-memory-hot_idle / durable-hot split that would be invisible to BOTH reuse and the TTL reaper. A sustained nonzero rate points at runtime-store write problems.",
-}, []string{"image"})
 
 // --- Observation helpers ---
 
@@ -328,30 +323,11 @@ func observeDrainTotalDuration(d time.Duration) {
 	workerDrainTotalDurationHistogram.Observe(d.Seconds())
 }
 
-// setJanitorIsLeader records whether this replica currently holds the janitor
-// leader lease (1) or not (0).
-func setJanitorIsLeader(isLeader bool) {
-	if isLeader {
-		janitorIsLeaderGauge.Set(1)
-		return
-	}
-	janitorIsLeaderGauge.Set(0)
-}
-
 // observeHotIdleReapRun stamps the last-successful-hot-idle-reap gauge with the
-// given time. Called by both the leader janitor and the per-CP fallback reaper.
+// given time. Called by both the leader janitor and the per-CP fallback reaper
+// (both untagged), so it lives here rather than in the kubernetes-tagged file.
 func observeHotIdleReapRun(t time.Time) {
 	hotIdleReapLastRunGauge.Set(float64(t.Unix()))
-}
-
-// observeHotIdlePersistFailure increments the hot_idle persist-failure
-// counter. Empty image falls back to "unknown".
-func observeHotIdlePersistFailure(image string) {
-	img := strings.TrimSpace(image)
-	if img == "" {
-		img = "unknown"
-	}
-	workerHotIdlePersistFailuresCounter.WithLabelValues(img).Inc()
 }
 
 // observeHealthCheck records the result of one health-check probe.
